Move integration tests to where they belong
