+++ nucliadb_texts/src/writer.rs
@@ -260,100 +260,3 @@ impl TextWriterService {
         }
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use std::collections::HashMap;
-    use std::time::SystemTime;
-
-    use nucliadb_core::protos::prost_types::Timestamp;
-    use nucliadb_core::{protos, NodeResult};
-    use tantivy::collector::{Count, TopDocs};
-    use tantivy::query::{AllQuery, TermQuery};
-    use tempfile::TempDir;
-
-    use super::*;
-
-    fn create_resource(shard_id: String) -> Resource {
-        let resource_id = ResourceId {
-            shard_id: shard_id.to_string(),
-            uuid: "f56c58ac-b4f9-4d61-a077-ffccaadd0001".to_string(),
-        };
-
-        let now = SystemTime::now()
-            .duration_since(SystemTime::UNIX_EPOCH)
-            .unwrap();
-        let timestamp = Timestamp {
-            seconds: now.as_secs() as i64,
-            nanos: 0,
-        };
-
-        let metadata = protos::IndexMetadata {
-            created: Some(timestamp.clone()),
-            modified: Some(timestamp),
-        };
-
-        const DOC1_TI: &str = "This is the first document";
-        const DOC1_P1: &str = "This is the text of the second paragraph.";
-        const DOC1_P2: &str = "This should be enough to test the tantivy.";
-        const DOC1_P3: &str = "But I wanted to make it three anyway.";
-
-        let ti_title = protos::TextInformation {
-            text: DOC1_TI.to_string(),
-            labels: vec!["/l/mylabel".to_string()],
-        };
-
-        let ti_body = protos::TextInformation {
-            text: DOC1_P1.to_string() + DOC1_P2 + DOC1_P3,
-            labels: vec!["/f/body".to_string()],
-        };
-
-        let mut texts = HashMap::new();
-        texts.insert("title".to_string(), ti_title);
-        texts.insert("body".to_string(), ti_body);
-
-        Resource {
-            resource: Some(resource_id),
-            metadata: Some(metadata),
-            texts,
-            status: protos::resource::ResourceStatus::Processed as i32,
-            labels: vec![],
-            paragraphs: HashMap::new(),
-            paragraphs_to_delete: vec![],
-            sentences_to_delete: vec![],
-            relations_to_delete: vec![],
-            relations: vec![],
-            vectors: HashMap::default(),
-            vectors_to_delete: HashMap::default(),
-            shard_id,
-        }
-    }
-
-    #[test]
-    fn test_new_writer() -> NodeResult<()> {
-        let dir = TempDir::new().unwrap();
-        let fsc = TextConfig {
-            path: dir.path().join("texts"),
-        };
-
-        let mut field_writer_service = TextWriterService::start(&fsc).unwrap();
-        let resource1 = create_resource("shard1".to_string());
-        let _ = field_writer_service.set_resource(&resource1);
-        let _ = field_writer_service.set_resource(&resource1);
-
-        let reader = field_writer_service.index.reader()?;
-        let searcher = reader.searcher();
-
-        let query = TermQuery::new(
-            Term::from_field_text(field_writer_service.schema.text, "document"),
-            IndexRecordOption::Basic,
-        );
-
-        let (_top_docs, count) = searcher.search(&query, &(TopDocs::with_limit(2), Count))?;
-        assert_eq!(count, 1);
-
-        let (_top_docs, count) = searcher.search(&AllQuery, &(TopDocs::with_limit(10), Count))?;
-        assert_eq!(count, 2);
-        Ok(())
-    }
-}
